# internal/app/api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gamaj API",
    "version": "is.0.0.1",
    "description": "Go-native Gamaj panel API for dashboard, bot, and third-party client developers.\n\nAuthentication: the dashboard uses an opaque server-side session stored in the HttpOnly `gamaj_session` cookie. Session permissions are loaded from the database on every request, so permission changes take effect without logging out. Bots and external integrations should use personal `rk_...` API keys as `Authorization: Bearer <api_key>`. Legacy JWT creation at `POST /admin/token` and `POST /api/admin/token` remains available only for older external sales-bot-compatible clients.\n\nDo not automate username/password login for bots. API keys can be revoked independently and inherit the owner admin permissions, service scope, and limits.\n\nMutation behavior: user, service, node, config, and host mutations are transactional where applicable. Runtime changes are synchronized to nodes through gRPC and recorded for recovery."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Gamaj panel origin"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    },
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/myaccount/api-keys": {
      "get": {
        "tags": [
          "My Account"
        ],
        "summary": "List personal API keys",
        "description": "List API keys owned by the current admin.\n\nAuthentication: send `Authorization: Bearer <api_key>` for bots and third-party integrations. Create personal admin API keys from My Account -> API Keys; each key inherits the owner admin permissions and limits. The current dashboard JWT from `POST /api/admin/token` is also accepted for interactive dashboard sessions and for Swagger UI when opened from inside the panel.\n\nPermissions: Requires any active authenticated admin. The target account is always the current bearer principal.",
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "getMyaccountKeys"
      },
      "post": {
        "tags": [
          "My Account"
        ],
        "summary": "Create personal API key",
        "description": "Create a new admin API key for the current account.\n\nAuthentication: send `Authorization: Bearer <api_key>` for bots and third-party integrations. Create personal admin API keys from My Account -> API Keys; each key inherits the owner admin permissions and limits. The current dashboard JWT from `POST /api/admin/token` is also accepted for interactive dashboard sessions and for Swagger UI when opened from inside the panel.\n\nPermissions: Requires any active authenticated admin. The target account is always the current bearer principal.\n\nRequest body: send JSON unless this operation explicitly declares multipart/form-data. Unknown fields are generally ignored only when the underlying handler is compatibility-oriented; business-rule violations return 400/403/409.\n\nSide effects: this is a mutation. Database writes are transactional where applicable; runtime-affecting changes enqueue node operations or call the node controller as documented by the route.",
        "requestBody": {
          "$ref": "#/components/requestBodies/JSON"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JSON"
          }
        },
        "operationId": "postMyaccountKeys"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Gamaj admin API key or legacy JWT",
        "description": "Use a personal rk_ API key for bots and third-party integrations. Legacy JWTs from /admin/token remain accepted for older external sales-bot-compatible clients."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "gamaj_session",
        "description": "Opaque HttpOnly dashboard session cookie issued by POST /api/auth/login. It is managed by the browser and is not readable by dashboard JavaScript."
      }
    },
...
    "requestBodies": {
      "JSON": {
        "required": false,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Object"
            },
            "example": {
              "name": "example",
              "status": "active",
              "settings": {}
            }
          }
        },
        "description": "Generic JSON request body. See each operation description for required business fields and side effects."
      },
...
    },
    "responses": {
      "JSON": {
        "description": "JSON response. Exact object fields follow the dashboard-compatible Gamaj API response for the route.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AnyValue"
            },
            "example": {
              "id": 1,
              "status": "ok",
              "detail": "Operation completed"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "additionalProperties": true,
        "example": {
          "name": "example",
          "status": "active",
          "settings": {}
        }
      },
      "AnyValue": {
        "nullable": true,
        "example": {
          "id": 1,
          "status": "ok"
        },
        "oneOf": [
          {
            "type": "object",
            "additionalProperties": true
          },
          {
            "type": "array",
            "items": {}
          },
          {
            "type": "string"
          },
          {
            "type": "number"
          },
          {
            "type": "integer"
          },
          {
            "type": "boolean"
          }
        ]
      },
...
    }
  }
}